Cloud_Function/requirements.txt
requests
google-cloud-pubsub
